Move metadata loading to common module
